+++ app/src/main/java/woowacourse/movie/view/data/TableSize.kt
@@ -1,3 +1,3 @@
-package woowacourse.movie.domain
+package woowacourse.movie.view.data
 
 data class TableSize(val row: Int, val column: Int)
