+++ app/src/main/java/woowacourse/movie/view/mapper/ImageMapper.kt
@@ -1,18 +1,18 @@
 package woowacourse.movie.view.mapper
 
+import woowacourse.movie.R
 import woowacourse.movie.domain.Image
 import woowacourse.movie.view.data.ImageViewData
 
 object ImageMapper : Mapper<Image, ImageViewData> {
+    private val images: List<Pair<Image, ImageViewData>> =
+        listOf(Image(0) to ImageViewData(R.drawable.poster_harrypotter))
+
     override fun Image.toView(): ImageViewData {
-        return ImageViewData(
-            resource
-        )
+        return images.find { it.first == this }?.second ?: throw IllegalStateException()
     }
 
     override fun ImageViewData.toDomain(): Image {
-        return Image(
-            resource
-        )
+        return images.find { it.second == this }?.first ?: throw IllegalStateException()
     }
 }
