+++ domain/src/main/java/woowacourse/movie/domain/MovieMock.kt
@@ -1,11 +1,10 @@
 package woowacourse.movie.domain
 
-import woowacourse.movie.R
 import java.time.LocalDate
 
 object MovieMock {
     fun create(): Movie = Movie(
-        Image(R.drawable.poster_harrypotter),
+        Image(0),
         "해리 포터",
         DateRange(
             LocalDate.of(2024, 3, 1),
