+++ domain/src/main/java/woowacourse/movie/domain/ReservationDetail.kt
