+++ domain/src/main/java/woowacourse/movie/domain/seat/MovieSeatRow.kt
