+++ domain/src/main/java/woowacourse/movie/domain/seat/Seat.kt
