+++ domain/src/main/java/woowacourse/movie/domain/seat/SeatRank.kt
