+++ domain/src/main/java/woowacourse/movie/domain/seat/SeatRow.kt
