+++ domain/src/main/java/woowacourse/movie/domain/seat/Seats.kt
